app-dev-project/requirements.txt
fastapi==0.85.1
pydantic==1.10.2
email-validator==1.3.0
uvicorn==0.18.3
typing_extensions>=4.4
Flask==2.1.3
passlib==1.7.4
python-jose==3.3.0
python-dotenv==0.21.0
python-multipart==0.0.5
pytest==7.2.0
requests==2.28.1
asyncpg==0.27.0
a2wsgi==1.6.0
tortoise-orm==0.19.2
aerich==0.7.1
gunicorn==20.1.0
pytest-cov==4.0.0
black==22.10.0
isort==5.10.1
newspaper3k==0.2.8
celery==5.2.7
sentry-sdk==1.13.0
emails==0.6
